cdc: move scheduler and agent to package base
