+++ cdc/scheduler/base/info_provider_test.go
@@ -11,20 +11,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package scheduler
+package base
 
 import (
 	"testing"
 
 	"github.com/pingcap/tiflow/cdc/model"
+	"github.com/pingcap/tiflow/cdc/scheduler"
 	"github.com/pingcap/tiflow/cdc/scheduler/util"
 	"github.com/stretchr/testify/require"
 )
 
 // Asserts that BaseSchedulerDispatcher implements InfoProvider interface.
-var _ InfoProvider = (*BaseScheduleDispatcher)(nil)
+var _ scheduler.InfoProvider = (*ScheduleDispatcher)(nil)
 
-func injectSchedulerStateForInfoProviderTest(dispatcher *BaseScheduleDispatcher) {
+func injectSchedulerStateForInfoProviderTest(dispatcher *ScheduleDispatcher) {
 	dispatcher.captureStatus = map[model.CaptureID]*captureStatus{
 		"capture-1": {
 			SyncStatus:   captureSyncFinished,
