+++ cdc/scheduler/base/schedule_dispatcher_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package scheduler
+package base
 
 import (
 	"fmt"
@@ -19,6 +19,8 @@ import (
 
 	"github.com/pingcap/log"
 	"github.com/pingcap/tiflow/cdc/model"
+	"github.com/pingcap/tiflow/cdc/scheduler"
+	"github.com/pingcap/tiflow/cdc/scheduler/base/protocol"
 	"github.com/pingcap/tiflow/cdc/scheduler/util"
 	cdcContext "github.com/pingcap/tiflow/pkg/context"
 	"github.com/stretchr/testify/mock"
@@ -26,7 +28,7 @@ import (
 	"go.uber.org/zap"
 )
 
-var _ ScheduleDispatcherCommunicator = (*mockScheduleDispatcherCommunicator)(nil)
+var _ scheduler.ScheduleDispatcherCommunicator = (*mockScheduleDispatcherCommunicator)(nil)
 
 const (
 	defaultEpoch = "default-epoch"
@@ -61,7 +63,7 @@ func (m *mockScheduleDispatcherCommunicator) DispatchTable(
 	tableID model.TableID,
 	captureID model.CaptureID,
 	isDelete bool,
-	epoch model.ProcessorEpoch,
+	epoch protocol.ProcessorEpoch,
 ) (done bool, err error) {
 	if !m.isBenchmark {
 		log.Info("dispatch table called",
@@ -115,8 +117,8 @@ func TestDispatchTable(t *testing.T) {
 		cf1, "capture-2").Return(true, nil)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	dispatcher.OnAgentSyncTaskStatuses("capture-1", defaultEpoch, []model.TableID{}, []model.TableID{}, []model.TableID{})
@@ -129,8 +131,8 @@ func TestDispatchTable(t *testing.T) {
 		Return(false, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	communicator.Reset()
@@ -145,8 +147,8 @@ func TestDispatchTable(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 	require.NotEqual(t, 0, len(communicator.addTableRecords["capture-1"]))
 	require.NotEqual(t, 0, len(communicator.addTableRecords["capture-2"]))
@@ -159,8 +161,8 @@ func TestDispatchTable(t *testing.T) {
 	communicator.ExpectedCalls = nil
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 
 	communicator.AssertExpectations(t)
 
@@ -200,8 +202,8 @@ func TestSyncCaptures(t *testing.T) {
 
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1500, []model.TableID{1, 2, 3, 4, 5}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 
 	communicator.Reset()
 	communicator.On("Announce", mock.Anything, cf1, "capture-1").
@@ -210,14 +212,14 @@ func TestSyncCaptures(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1500, []model.TableID{1, 2, 3, 4, 5}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 
 	dispatcher.OnAgentSyncTaskStatuses("capture-1", defaultEpoch, []model.TableID{1, 2, 3}, []model.TableID{4, 5}, []model.TableID{6, 7})
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1500, []model.TableID{1, 2, 3, 4, 5}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 
 	communicator.Reset()
 	dispatcher.OnAgentFinishedTableOperation("capture-1", 4, defaultEpoch)
@@ -225,8 +227,8 @@ func TestSyncCaptures(t *testing.T) {
 	dispatcher.OnAgentSyncTaskStatuses("capture-2", defaultEpoch, []model.TableID(nil), []model.TableID(nil), []model.TableID(nil))
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1500, []model.TableID{1, 2, 3, 4, 5}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 
 	communicator.Reset()
 	dispatcher.OnAgentFinishedTableOperation("capture-1", 6, defaultEpoch)
@@ -255,8 +257,8 @@ func TestSyncUnknownCapture(t *testing.T) {
 	// We expect the `Sync` to be ignored.
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1500, []model.TableID{1, 2, 3, 4, 5}, mockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 }
 
 func TestRemoveTable(t *testing.T) {
@@ -307,8 +309,8 @@ func TestRemoveTable(t *testing.T) {
 		Return(false, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1500, []model.TableID{1, 2}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	communicator.Reset()
@@ -317,8 +319,8 @@ func TestRemoveTable(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1500, []model.TableID{1, 2}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	dispatcher.OnAgentFinishedTableOperation("capture-1", 3, defaultEpoch)
@@ -379,8 +381,8 @@ func TestCaptureGone(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1500, []model.TableID{1, 2, 3}, mockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 }
 
@@ -427,8 +429,8 @@ func TestCaptureRestarts(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1500, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 }
 
@@ -486,8 +488,8 @@ func TestCaptureGoneWhileMovingTable(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, mockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	delete(mockCaptureInfos, "capture-2")
@@ -501,8 +503,8 @@ func TestCaptureGoneWhileMovingTable(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, mockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 }
 
@@ -562,8 +564,8 @@ func TestRebalance(t *testing.T) {
 		Return(false, nil)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3, 4, 5, 6}, mockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 	communicator.AssertNumberOfCalls(t, "DispatchTable", 1)
 
@@ -573,8 +575,8 @@ func TestRebalance(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3, 4, 5, 6}, mockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertNumberOfCalls(t, "DispatchTable", 2)
 	communicator.AssertExpectations(t)
 }
@@ -706,8 +708,8 @@ func TestIgnoreUnsyncedCaptures(t *testing.T) {
 	dispatcher.OnAgentCheckpoint("capture-2", 1000, 1000)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3, 4, 5, 6}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 
 	communicator.Reset()
 	dispatcher.OnAgentSyncTaskStatuses("capture-2", defaultEpoch, []model.TableID{2, 4, 6}, []model.TableID{}, []model.TableID{})
@@ -752,16 +754,16 @@ func TestRebalanceWhileAddingTable(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3, 4, 5, 6, 7}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	dispatcher.Rebalance()
 	communicator.Reset()
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3, 4, 5, 6, 7}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	dispatcher.OnAgentFinishedTableOperation("capture-2", model.TableID(7), defaultEpoch)
@@ -771,8 +773,8 @@ func TestRebalanceWhileAddingTable(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3, 4, 5, 6, 7}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertNumberOfCalls(t, "DispatchTable", 2)
 	communicator.AssertExpectations(t)
 }
@@ -814,14 +816,14 @@ func TestManualMoveTableWhileAddingTable(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 
 	dispatcher.MoveTable(1, "capture-1")
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	dispatcher.OnAgentFinishedTableOperation("capture-2", 1, defaultEpoch)
@@ -831,8 +833,8 @@ func TestManualMoveTableWhileAddingTable(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	dispatcher.OnAgentFinishedTableOperation("capture-2", 1, defaultEpoch)
@@ -842,8 +844,8 @@ func TestManualMoveTableWhileAddingTable(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 }
 
@@ -881,8 +883,8 @@ func TestAutoRebalanceOnCaptureOnline(t *testing.T) {
 		cf1, "capture-2").Return(true, nil)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, captureList)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	dispatcher.OnAgentSyncTaskStatuses("capture-1", defaultEpoch, []model.TableID{}, []model.TableID{}, []model.TableID{})
@@ -900,8 +902,8 @@ func TestAutoRebalanceOnCaptureOnline(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, captureList)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 	require.NotEqual(t, 0, len(communicator.addTableRecords["capture-1"]))
 	require.NotEqual(t, 0, len(communicator.addTableRecords["capture-2"]))
@@ -914,8 +916,8 @@ func TestAutoRebalanceOnCaptureOnline(t *testing.T) {
 	communicator.ExpectedCalls = nil
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, captureList)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 
 	communicator.AssertExpectations(t)
 
@@ -928,16 +930,16 @@ func TestAutoRebalanceOnCaptureOnline(t *testing.T) {
 		cf1, "capture-3").Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, captureList)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	communicator.ExpectedCalls = nil
 	dispatcher.OnAgentSyncTaskStatuses("capture-3", defaultEpoch, []model.TableID{}, []model.TableID{}, []model.TableID{})
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, captureList)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	for captureID, tables := range communicator.addTableRecords {
@@ -955,8 +957,8 @@ func TestAutoRebalanceOnCaptureOnline(t *testing.T) {
 	})
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, captureList)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 
 	removedTableID := communicator.removeTableRecords[removeTableFromCapture][0]
@@ -970,8 +972,8 @@ func TestAutoRebalanceOnCaptureOnline(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1000, []model.TableID{1, 2, 3}, captureList)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	communicator.AssertExpectations(t)
 }
 
@@ -1012,15 +1014,15 @@ func TestInvalidFinishedTableOperation(t *testing.T) {
 		Return(true, nil)
 	checkpointTs, resolvedTs, err := dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 
 	// Invalid epoch
 	dispatcher.OnAgentFinishedTableOperation("capture-2", model.TableID(1), "invalid-epoch")
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	record, ok := dispatcher.tables.GetTableRecord(model.TableID(1))
 	require.True(t, ok)
 	require.Equal(t, record.Status, util.AddingTable)
@@ -1029,8 +1031,8 @@ func TestInvalidFinishedTableOperation(t *testing.T) {
 	dispatcher.OnAgentFinishedTableOperation("capture-invalid", model.TableID(1), defaultEpoch)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	record, ok = dispatcher.tables.GetTableRecord(model.TableID(1))
 	require.True(t, ok)
 	require.Equal(t, record.Status, util.AddingTable)
@@ -1039,8 +1041,8 @@ func TestInvalidFinishedTableOperation(t *testing.T) {
 	dispatcher.OnAgentFinishedTableOperation("capture-1", model.TableID(999), defaultEpoch)
 	checkpointTs, resolvedTs, err = dispatcher.Tick(ctx, 1300, []model.TableID{1, 2, 3}, defaultMockCaptureInfos)
 	require.NoError(t, err)
-	require.Equal(t, CheckpointCannotProceed, checkpointTs)
-	require.Equal(t, CheckpointCannotProceed, resolvedTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, checkpointTs)
+	require.Equal(t, scheduler.CheckpointCannotProceed, resolvedTs)
 	record, ok = dispatcher.tables.GetTableRecord(model.TableID(1))
 	require.True(t, ok)
 	require.Equal(t, record.Status, util.AddingTable)
